requirements.txt
torch
torchvision
accelerate
bitsandbytes
transformers
sentence_transformers
InstructorEmbedding
llama-cpp-python
beautifulsoup4
openai
tiktoken
pytesseract
pdf2image
unstructured
python-dotenv
faiss-cpu
promptlayer
langchain
llama_index